requirements.txt
Flask==2.2.5
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
oracledb==3.0.0
python-dotenv==1.1.0
SQLAlchemy==2.0.27
Werkzeug==2.2.3